Moved constant package from option/command to option/.  Tests are very broken.
