+++ internal/security/secrets/option/command/cache/cache.go
@@ -23,8 +23,8 @@ import (
 	"path/filepath"
 
 	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option"
-	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/command/constant"
 	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/command/generate"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/constant"
 
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/logger"
 )
