+++ internal/security/secrets/option/command/generate/flagset.go
@@ -17,7 +17,7 @@ package generate
 import (
 	"flag"
 
-	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/command/constant"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/constant"
 )
 
 type FlagSet struct {
