+++ internal/security/secrets/option/command/generate/generate.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/edgexfoundry/edgex-go/internal/pkg/config"
 	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option"
-	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/command/constant"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/constant"
 
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/logger"
 )
