+++ internal/security/secrets/option/command/legacy/legacy.go
@@ -18,7 +18,7 @@ import (
 	"flag"
 
 	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option"
-	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/command/constant"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/constant"
 )
 
 type Command struct {
