+++ internal/security/secrets/option/optionHelpers.go
@@ -19,7 +19,7 @@ package option
 import (
 	"errors"
 	"fmt"
-	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/command/constant"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets/option/constant"
 	"io"
 	"io/ioutil"
 	"os"
